Move GitRepo to git
